ecosense-ai/backend/requirements.txt
# ===========================================
# EcoSense AI — Python Dependencies
# ===========================================

# Django Core
Django==4.2.7
gunicorn==21.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7

# Celery + Redis
celery==5.3.4
celery[redis]==5.3.4
redis==4.5.4
django-celery-beat==2.5.0

# Geospatial
geopandas==0.14.0
shapely==2.0.2


# AI / ML
langchain==0.0.335
openai==1.3.3
xgboost==2.0.1
scikit-learn==1.3.2

# Report Generation
WeasyPrint==60.1
python-docx==1.1.0

# Blockchain
web3==6.11.1

# AWS
boto3==1.29.0

# Imaging
Pillow==10.1.0

# Messaging
africastalking==1.2.3
twilio==8.10.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
